changed id type from int to long

+++ src/ParallelArrayObject.java
@@ -1,20 +1,20 @@
 public class ParallelArrayObject {
 
-    private int id;
+    private long id;
     private double total;
     private int grupo;
 
-    public ParallelArrayObject(int _id, double _total, int _grupo) {
+    public ParallelArrayObject(long _id, double _total, int _grupo) {
         id = _id;
         total = _total;
         grupo = _grupo;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int _id) {
+    public void setId(long _id) {
         id = _id;
     }
